Refactor SkyModel to support telstate
Instead of SkyModel doing all the work itself, it's now a base class for
KatpointSkyModel. The work of actually constructing the catalogue is
moved out into a free function so that KatpointSkyModel can be directly
constructed with a catalogue for unit testing.

The file-opening function now takes an URL rather than a filename, and
supports file URLs for local catalogue files or a specific redis://
scheme to get data from telstate. It repeats a lot of the information
that could be found from other parameters when opening a katdal file,
but this way it's orthogonal and so one can mix and match sky models for
debugging; and it's not expected that in normal use these URLs will be
constructed by katsdpcontroller rather than typed in.

There are now better unit tests. add_to_image was removed because it's
not used anywhere any more and would have been a pain to test.