Solve for BCROSS_DIODE
This commit allows to the pipeline to solve for the
HV phase as a function of channel using autocorrelation data if the
noise diode is fired. The solution is formed by :
* averaging HV with the conjugate of VH
* taking the mean phase of these averaged visibilities per antenna
* the solutions on the second polarisation axis are set to zero.

The HV phase is solved for after applying parallel hand calibration solutions
and the KCROSS_DIODE solutions. These solutions do not commute and must proceed in
the given order.

The way in which HV phase and delay solutions are applied to the data is also modified.
Only the solution at the chosen reference antenna is used to correct the data. This is to
ensure that small differences between the noise-diode calibration and celestial source based
calibration at a given antenna remain consistent between separate observations with the same
reference antenna