Add tests for new BCROSS_DIODE product
The tests checks that if the noise diode is fired while observing a target tagged
'bfcal' a BCROSS_DIODE solution is produced and has the expected values.
The produced BCROSS_DIODE should be equal to the phase difference between the two polarisations at the
reference antenna in the first polarisation axis and zero in the second axis.

In order to compare the BCROSS_DIODE solution with the phase difference between antennas
it is necessary to interpolate across NaN's in the solution (caused by flagging in the pipeline).

Interpolation of solutions across NaN's is now also be used to verify the B (bandpass) solutions
produced by the pipeline.