Pass mfimage_params and work out which SN tables to export
The AIPS SN tables produced by MFImage are cumulative with respect
to the separte phase only and amp+phase self cal rounds. The SN
table version number corresponding to the full set of phase only
self-calibration is equal to the value of the MFImage parameter
maxPSCLoop and the version number of amp+phase self calibration
solutions is maxPSCLoop + maxASCLoop.