Add immutable telstate keys
The keys added in this commit are all derived from the katdal_adaptor
and are those described in JIRA SR-1742. The pol_ordering
is derived from the static MeerKAT pol to AIPS corr ID defined
in katdal_adapter.py. katdal_adapter.py has limited-to-nonexistent
support for anything other than dual-pol UV data (with STOKES=-5)
so this commit dropps support for single-pol data in the solution
export.