Plot model spectrum in report
For plots of the spectra of gain calibrated targets include the spectrum
of the model of the source. This helps to validate that calibration has
corrected scaled the visibilites to the source model.

For sources with multiple sources in model only
plot the spectrum of the first source which is assumed to be the brightest and
dominant source in the field.