Change namespace of the 'refant' in telstate
Store refant in the 'cal_' namespace instead of the 'cbid_cal_' namespace.
As the refant is set only once per sub-array and not once per capture block
it logically should be stored in the same namespace as all the other per sub-array
cal parameters. This changle also allows scripts such as `bf_phaseup.py` to read the refant
which will be required to apply the appropriate 'BCROSS_DIODE' and 'KCROSS_DIODe' solutions.

Additionally the 'refant' key is replaced/superceded by the 'refant_index' key, which now
only stores the index number of the reference antenna (instead of the whole antenna
description string)