Add refant to each new cbid namespace in telstate
Currently the refant is added to the `cbid_cal` namespace in telstate.
As the refant is only set once for the entire subarray this means that currently
the refant key is only present in the first capture block (cbid) of the subarray.
This PR checks whether the `refant` key is present in the `cbid_cal` namespace
for every calibrator scan, if the key doesn't exist then it is set to the refant value.

Additionally the `refant` key no longer stores a full antenna description string, but instead
simply stores the name of the reference antenna.