Create a pipeline_factory function to choose a pipeline implementation.
Modify the various pipeline classes so that they all inherit from an interface
class (with an execute method) called Pipeline, and a heirachy of implementation
classes all of which inherit from a class called PipelineImplentation.

The pipeline_factory function decides, based upon a tag, and investgating the
data type of the input datawhich pipeline instance to instantiate. At present
available tags are 'online', 'offline' (for the versions in previous commits
in the PR) and 'export_continuum' (which is currently only used for testing).
The only available data type is a katdal DataSet.

Cleanup of files on the aipsdisk after termination is now done by requiring that
each concrete pipeline class implements a context manager which runs the
cleanup function on exit.